requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-allauth==65.3.1
idna==3.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0